requirements.txt
langsmith>=0.0.52
langchain-qdrant
langchain-core>=0.0.2
aiomonitor
python-dotenv>=0.19.0
mysql-connector-python
openai>=1.1.1
fastapi>=0.68.0
uvicorn>=0.15.0
langchain>=0.0.200
langchain-openai>=0.0.2
sentence-transformers
pypdf
nest-asyncio>=1.5.1
IPython
PyPDF2
python-docx
beautifulsoup4
python-multipart>=0.0.5
aiosqlite>=0.17.0
qdrant-client>=1.1.1
asyncio>=3.4.3
numpy>=1.21.0
scikit-learn>=0.24.2
fastapi-cache2>=0.2.1
aioredis>=2.0.1
redis>=4.5.1
langchain-community>=0.0.2